requirements.txt
flask==2.1.1
pillow==10.4.0
gunicorn==23.0.0
requests==2.32.3
tensorflow==2.7.0
numpy ==1.26.4
jinja2==3.1.6
markupsafe==2.1.1
werkzeug==2.1.1
keras==3.9.2